LayoutTests/platform/chromium/http/tests/uri/resolve-encoding-relative-expected.txt

Test how non-ASCII characters are encoded in relative URLs.

1 2 3 4 5
1. PASS
2. FAIL: http://127.0.0.1:8000/uri/resolve-encoding-relative.html#фрагмент
3. PASS
4. PASS
5. FAIL: http://127.0.0.1:8000/%D0%BF%D1%83%D1%82%D1%8C?%E7%E0%EF%F0%EE%F1#фрагмент
